Happy Number commit on 25-09-2026

+++ loops/PrimeFirstLastDigit.java
@@ -39,7 +39,7 @@ public class PrimeFirstLastDigit {
 		if(n<2) {
 			return false;
 		}
-		for(int i=2;i*i<=n;i++) {
+		for(int i=2;(long)i*i<=n;i++) {
 			if(n%i==0) {
 				flag=false;
 				break;
